Update to introduce an example script and restoring UnityCanvasTransport.cs script back to its original
