+++ Examples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 45e132fef99756344a88f9ed89bf7112
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
