+++ Examples/AnchorCanvasExample.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ad9b6af9286269488bdf459d596a17c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
