+++ Examples/AnchorCanvasExample/ExampleCanvasScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 93436d97ee043924ebb50a190563bc55
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
