+++ Examples/AnchorCanvasExample/ExampleUnityCanvasTransport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d50a4ee8395312439511422bc994028
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
