+++ Examples/AnchorCanvasExample/QRCodeRectAnimationAssets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f651f4b138286a42b16024b70cb833e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
