+++ Examples/AnchorCanvasExample/QRCodeRectAnimationAssets/NormalState.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f8a183d50145a643a2ca48fab0cda09
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
