+++ Examples/AnchorCanvasExample/QRCodeRectAnimationAssets/QRCodeRectScaler.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 044c1daa746b9a348ae0e5d392ca0b74
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
