+++ Examples/AnchorCanvasExample/QRCodeRectAnimationAssets/ZoomIn.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83d24b6ea54bfc742a9d151898bffc94
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
