+++ Examples/AnchorCanvasExample/QRCodeRectAnimationAssets/ZoomOut.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2fa0e756b3019d74a82f7ac507de98f7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
