+++ Examples/AnchorUIToolkitExample.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: feb209a201f3f5d4e8aa7b4e7c0b2d06
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
